Working i386 port.

+++ kernel/include/e820.h
@@ -11,7 +11,7 @@ typedef struct {
     uint32_t unused;
 } __attribute__((packed)) e820_entry_t;
 
-extern size_t memory_size;
+extern uint64_t memory_size;
 extern e820_entry_t e820_map[256];
 
 void init_e820(void);
